aerc: init colors
There are a lot more options available, as can be seen at
aerc-stylesets(7). The colorize filter colors could theoretically be
configured as well, but as of now, home-manager doesn't allow that, so
it needs to be postponed.

There are also other style settings that could be changed such as
boldness of text, but are not in order to leave the configuration to the
users.